requirements.txt
Flask>=2.0
requests
flask-cors
python-dotenv
